langgraph_agent.py
# backend/langgraph_agent.py
import os
from dotenv import load_dotenv
from rag_indexer import query_rag
load_dotenv(os.path.join(os.path.dirname(__file__), "..", "config.env"))

# This is an example skeleton. Your setup for LangGraph may vary.
try:
    from langgraph import LangGraph, Tool, Agent
    LANGGRAPH_AVAILABLE = True
except Exception:
    LANGGRAPH_AVAILABLE = False

def run_browser_agent(question: str):
    """
    Use LangGraph to run a browser search + RAG fallback.
    This is a minimal example. If LangGraph is not installed, this will return an informative message.
    """
    if not LANGGRAPH_AVAILABLE:
        return {"answer": "LangGraph not installed - install 'langgraph' to enable web-agent.", "sources": []}

    # Create a simple tool that runs a web search. Replace with a real implementation if available.
    def web_search_tool(q):
        # Very simple: Use an external web search API if available or return a placeholder.
        # For hackathon, you may implement an http GET to a search API here.
        return f"Search results placeholder for query: {q}"

    web_tool = Tool(name="web_search", func=web_search_tool, description="Run web search")

    # Compose an agent that can call web_search, or fall back to RAG
    agent = Agent(tools=[web_tool])

    # First try RAG
    rag_result = query_rag(question)
    # If RAG gives a short/low-confidence answer, call web_search
    # (No explicit confidence from RetrievalQA here; adapt as needed)
    if rag_result.get("answer") and len(rag_result.get("answer")) > 20:
        return rag_result

    # Else call web tool
    web_out = agent.run(question)
    # merge responses
    return {"answer": f"Web agent output:\n{web_out}\n\nRAG fallback:\n{rag_result.get('answer')}", "sources": rag_result.get("sources")}
